+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   code.
 * Added optimization of restrict over unwrap when the resulting type is known
   and a proper TupleType.
+* Fixed type inference to favor Boolean over TrueClass and FalseClass
 
 # 0.16.3 - 2014-07-25